requirements.txt
flask
flask_pymongo
PyJWT
flask_cors
bcrypt
pymongo
dnspython